+++ src/ExpressionCache.ts
@@ -16,7 +16,17 @@ export default class ExpressionCache {
     this.parentPage = options.parentPage
   }
 
-  evaluate<TResult> (expression: string, page: Page) {
+  /** Helper function to determine if a given value looks like it should be
+   * rendered.
+   */
+  appearsRenderable (value: any) {
+    if (value === null) { return false }
+    if (typeof value === 'object') { return true }
+    if (typeof value === 'string' && new RegExp('^#[0-9a-z\-_/]+', 'i').exec(value)) { return true }
+    return false
+  }
+
+  evaluate<TResult> (expression: string, page: Page | null = null) {
     const cacheKey = JSON.stringify({
       parent: {
         path: this.parentPage.file.path,
@@ -24,9 +34,9 @@ export default class ExpressionCache {
         size: this.parentPage.file.size
       },
       page: {
-        path: page.file.path,
-        mtime: page.file.mtime.toMillis(),
-        size: page.file.size
+        path: page?.file.path,
+        mtime: page?.file.mtime.toMillis(),
+        size: page?.file.size
       },
       expression,
       operation: 'evaluate'
@@ -46,7 +56,7 @@ export default class ExpressionCache {
     })
   }
 
-  async renderFieldValue (expression: string, page: Page): Promise<string | null> {
+  async renderExpression (expression: string, page: Page | null = null): Promise<string | null> {
     const cacheKey = JSON.stringify({
       parent: {
         path: this.parentPage.file.path,
@@ -54,22 +64,45 @@ export default class ExpressionCache {
         size: this.parentPage.file.size
       },
       page: {
-        path: page.file.path,
-        mtime: page.file.mtime.toMillis(),
-        size: page.file.size
+        path: page?.file.path,
+        mtime: page?.file.mtime.toMillis(),
+        size: page?.file.size
       },
       expression,
-      operation: 'renderFieldValue'
+      operation: 'renderExpression'
     })
 
     return this._cache.fetch(cacheKey, async () => {
       const value = this.evaluate(expression, page)
       if (!value) { return null }
 
+      return await this.renderFieldValue(value, page)
+    })
+  }
+
+  async renderFieldValue (value: any, page: Page | null = null): Promise<string | null> {
+    const cacheKey = JSON.stringify({
+      parent: {
+        path: this.parentPage.file.path,
+        mtime: this.parentPage.file.mtime.toMillis(),
+        size: this.parentPage.file.size
+      },
+      page: {
+        path: page?.file.path,
+        mtime: page?.file.mtime.toMillis(),
+        size: page?.file.size
+      },
+      value,
+      operation: 'renderFieldValue'
+    })
+
+    return this._cache.fetch(cacheKey, async () => {
+      if (!value) { return null }
+
       const temp = document.createElement('div')
-      await this.api.renderValue(value, temp, this.component, page.file.path, true)
+      await this.api.renderValue(value, temp, this.component, page?.file.path || this.parentPage.file.path, true)
 
       return temp.innerHTML
-      })
+    })
   }
 }